pyproject.toml
[project]
name = "mcpserverlangchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "langchain-groq>=0.3.2",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "mcp>=1.9.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
